+++ app/controllers/concerns/field_to_resource_lookup.rb
@@ -1,0 +1,8 @@
+module FieldToResourceLookup
+  FIELD_TO_RESOURCE_LOOKUP = {
+    contacts: {
+      avatar: :primary_picture,
+      square_avatar: :primary_picture
+    }
+  }.freeze
+end