+++ app/services/contact/filter/name.rb
@@ -1,5 +1,0 @@
-class Contact::Filter::Name < Contact::Filter::Base
-  def execute_query(contacts, filters, _account_lists)
-    contacts.where('lower(contacts.name) like ?', "%#{filters[:name].downcase}%")
-  end
-end