+++ spec/acceptance/api/v2/appeals/contacts_spec.rb
@@ -23,6 +23,7 @@ resource 'Appeals > Contacts' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting
@@ -54,7 +55,6 @@ resource 'Appeals > Contacts' do
       suggested_changes
       tag_list
       timezone
-      total_donations
       uncompleted_tasks_count
       updated_at
       updated_in_db_at