+++ spec/acceptance/api/v2/contacts/referrers_spec.rb
@@ -22,6 +22,7 @@ resource 'Contacts > Referrers' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting
@@ -53,7 +54,6 @@ resource 'Contacts > Referrers' do
       suggested_changes
       tag_list
       timezone
-      total_donations
       uncompleted_tasks_count
       updated_at
       updated_in_db_at