+++ spec/acceptance/api/v2/contacts_spec.rb
@@ -59,6 +59,7 @@ resource 'Contacts' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting
@@ -90,7 +91,6 @@ resource 'Contacts' do
       suggested_changes
       tag_list
       timezone
-      total_donations
       uncompleted_tasks_count
       updated_at
       updated_in_db_at
@@ -185,6 +185,8 @@ resource 'Contacts' do
       end
     end
 
+    let(:additional_attributes) { %w(lifetime_donations) }
+
     get '/api/v2/contacts/:id' do
       doc_helper.insert_documentation_for(action: :show, context: self)
       with_options scope: :relationships do
@@ -196,7 +198,7 @@ resource 'Contacts' do
         explanation doc_helper.description_for(:show)
         do_request
 
-        check_resource(additional_keys)
+        check_resource(additional_keys, additional_attributes)
         expect(resource_object['name']).to eq contact.name
         expect(response_status).to eq 200
       end